Tests/LibWeb/Text/expected/Fetch/stale-link-fetch-body-callback.txt

Final body color: rgb(0, 0, 255)
